[#3] Creational business logic completed
1. Factory pattern implemented to create productUnits, boxes and pallets in certain order.
2. (Temporarly) implemented Distributed Cache using memory cache, but prefer to use Redis in future.
3. Changed DateOnly type to DateTime because have some serialization mistakes.
4. (Temporarly) created Response and Request models for ActiveProduction status of application.

+++ ProductionAccounting.Core/Entities/Invoice.cs
@@ -6,7 +6,7 @@
 		public string Sender {  get; set; }
 		public string Recipient {  get; set; }
 		public DateOnly Date { get; set; }
-		public double TotalCost { get; set; }
+		public decimal TotalCost { get; set; }
 
 		public IEnumerable<Box>? Boxes { get; set; }
 		public IEnumerable<Pallet>? Pallets { get; set; }
